feat: summarize differences with RCDB
