+++ report/beam-energy-mon.yaml
@@ -1,0 +1,39 @@
+# hard-coded beam energies from bin/slurm-mon12.sh
+beamEnergy:
+  A:
+    - runnum: [ 3861, 5673 ]
+      value: 10.6
+    - runnum: [ 6608, 6783 ]
+      value: 10.199
+
+  K:
+    - runnum: [ 5674, 5870 ]
+      value: 7.546
+    - runnum: [ 5871, 6000 ]
+      value: 6.535
+
+  F:
+    - runnum: [ 11620, 11657 ] # not sure if actually RG-F
+      value: 2.182
+    - runnum: [ 11658, 12283 ]
+      value: 10.389
+    - runnum: [ 12389, 12444 ]
+      value: 2.182
+    - runnum: [ 12445, 12951 ]
+      value: 10.389
+
+  M:
+    - runnum: [ 15013, 15490 ]
+      value: 5.98636
+    - runnum: [ 15533, 15727 ]
+      value: 2.07052
+    - runnum: [ 15728, 15784 ]
+      value: 4.02962
+    - runnum: [ 15787, 15884 ]
+      value: 5.98636
+
+  C:
+    - runnum: [ 16010, 16079 ] # not sure if actually RG-C
+      value: 2.22
+    - runnum: [ 16084, 20000 ]
+      value: 10.54
